+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/repository/SearchFeedListRepository.kt
@@ -6,8 +6,10 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.dnd_9th_3_android.gooding.data.SearchFeedListPagingSource
 import com.dnd_9th_3_android.gooding.data.SearchFeedListRemoteDataSource
+import com.dnd_9th_3_android.gooding.data.model.FeedData
 import com.dnd_9th_3_android.gooding.data.model.PopularKeywordListResponse
 import com.dnd_9th_3_android.gooding.data.model.RecentKeywordListResponse
+import com.dnd_9th_3_android.gooding.data.model.RequestUploadFeed
 import com.dnd_9th_3_android.gooding.data.model.SearchFeedData
 import com.dnd_9th_3_android.gooding.data.model.toEntity
 import kotlinx.coroutines.flow.Flow
@@ -24,42 +26,4 @@ interface SearchFeedRepository {
     suspend fun addRecentKeywordList()
 
     suspend fun deleteRecentKeywordList()
-}
-
-class SearchFeedRepositoryImpl @Inject constructor(
-    private val searchFeedListRemoteDataSource: SearchFeedListRemoteDataSource
-): SearchFeedRepository {
-    override fun searchFeedRepository(query: String): Flow<PagingData<SearchFeedData>> {
-        return Pager(
-            config = PagingConfig(
-                pageSize = 20,
-                initialLoadSize = 20
-            ),
-            pagingSourceFactory = {
-                SearchFeedListPagingSource(searchFeedListRemoteDataSource, query)
-            }
-        )
-            .flow
-            .map { feedData ->
-                feedData.map {
-                    it.toEntity()
-                }
-            }
-    }
-
-    override suspend fun getRecentKeywordList(): RecentKeywordListResponse {
-        return searchFeedListRemoteDataSource.getRecentKeywordList()
-    }
-
-    override suspend fun getPopularKeywordList(): PopularKeywordListResponse {
-        return searchFeedListRemoteDataSource.getPopularKeywordList()
-    }
-
-    override suspend fun addRecentKeywordList() {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun deleteRecentKeywordList() {
-        TODO("Not yet implemented")
-    }
 }
